+++ temba/campaigns/tests.py
@@ -24,10 +24,10 @@ class CampaignTest(TembaTest):
     def setUp(self):
         super().setUp()
 
-        self.farmer1 = self.create_contact("Rob Jasper", "+250788111111")
-        self.farmer2 = self.create_contact("Mike Gordon", "+250788222222", language="spa")
+        self.farmer1 = self.create_contact("Rob Jasper", phone="+250788111111")
+        self.farmer2 = self.create_contact("Mike Gordon", phone="+250788222222", language="spa")
 
-        self.nonfarmer = self.create_contact("Trey Anastasio", "+250788333333")
+        self.nonfarmer = self.create_contact("Trey Anastasio", phone="+250788333333")
         self.farmers = self.create_group("Farmers", [self.farmer1, self.farmer2])
 
         self.reminder_flow = self.create_flow(name="Reminder Flow")
@@ -1200,7 +1200,7 @@ class CampaignTest(TembaTest):
         self.assertEqual(str(planting_reminder), 'Event[relative_to=planting_date, offset=0, flow="Reminder Flow"]')
 
         # schedule our reminders
-        EventFire.update_campaign_events(campaign)
+        campaign.schedule_events_async()
 
         # we should have queued a scheduling task to mailroom
         self.assertEqual(
@@ -1218,7 +1218,7 @@ class CampaignTest(TembaTest):
         # if any event fires already exist, scheduling will trigger cloning the event
         EventFire.objects.create(event=planting_reminder, contact=self.farmer1, scheduled=timezone.now(), fired=None)
 
-        EventFire.update_campaign_events(campaign)
+        campaign.schedule_events_async()
 
         planting_reminder.refresh_from_db()
         self.assertFalse(planting_reminder.is_active)
@@ -1361,7 +1361,7 @@ class CampaignTest(TembaTest):
         )
 
         # create a contact not in the group, but with a field value
-        anna = self.create_contact("Anna", urn="tel:+250788333333", fields={"planting_date": "09-10-2020 12:30"})
+        anna = self.create_contact("Anna", phone="+250788444444", fields={"planting_date": "09-10-2020 12:30"})
 
         # no contacts in our dynamic group yet, so no event fires
         self.assertEqual(EventFire.objects.filter(event=event).count(), 0)