+++ temba/campaigns/views.py
@@ -15,7 +15,7 @@ from temba.utils.fields import CompletionTextarea, InputWidget, SelectWidget
 from temba.utils.views import BulkActionMixin
 from temba.values.constants import Value
 
-from .models import Campaign, CampaignEvent, EventFire
+from .models import Campaign, CampaignEvent
 
 
 class UpdateCampaignForm(forms.ModelForm):
@@ -83,7 +83,7 @@ class CampaignCRUDL(SmartCRUDL):
 
             # if our group changed, create our new fires
             if new_group != previous_group:
-                EventFire.update_campaign_events(self.object)
+                self.object.schedule_events_async()
 
             response = self.render_to_response(
                 self.get_context_data(