+++ temba/channels/types/telegram/tests.py
@@ -80,7 +80,7 @@ class TelegramTypeTest(TembaTest):
             response.context["form"].errors["auth_token"][0],
         )
 
-        contact = self.create_contact("Telegram User", urn=URN.from_telegram("1234"))
+        contact = self.create_contact("Telegram User", urns=[URN.from_telegram("1234")])
 
         # make sure we our telegram channel satisfies as a send channel
         response = self.client.get(reverse("contacts.contact_read", args=[contact.uuid]))